Module Span.For_human

type sign = [

| `Pos

| `Neg

]

type view = private {

sign : sign;

days : int;

hours : int;

minutes : int;

seconds : int;

ns : int;

}

type error = [

| `Invalid_days of int

| `Invalid_hours of int

| `Invalid_minutes of int

| `Invalid_seconds of int

| `Invalid_ns of int

]

type error_f = [

| `Invalid_days_f of float

| `Invalid_hours_f of float

| `Invalid_minutes_f of float

| `Invalid_seconds_f of float

| `Invalid_ns of int

]

exception Error_exn of error

exception Error_f_exn of error_f

val make : ?⁠sign:sign -> ?⁠days:int -> ?⁠hours:int -> ?⁠minutes:int -> ?⁠seconds:int -> ?⁠ns:int -> unit -> (t, error) Stdlib.result

sign defaults to `Pos.

Returns Error if any of the arguments are negative.

val make_exn : ?⁠sign:sign -> ?⁠days:int -> ?⁠hours:int -> ?⁠minutes:int -> ?⁠seconds:int -> ?⁠ns:int -> unit -> t

raises Error_exn

if make fails

val make_frac : ?⁠sign:sign -> ?⁠days:float -> ?⁠hours:float -> ?⁠minutes:float -> ?⁠seconds:float -> ?⁠ns:int -> unit -> (t, error_f) Stdlib.result

sign defaults to `Pos.

Returns Error if any of the arguments are negative.

val make_frac_exn : ?⁠sign:sign -> ?⁠days:float -> ?⁠hours:float -> ?⁠minutes:float -> ?⁠seconds:float -> ?⁠ns:int -> unit -> t

raises Error_exn

if make fails

val view : t -> view

val pp : Stdlib.Format.formatter -> t -> unit

val to_string : t -> string