Module Timedesc.Zoneless

type zoneless

type error_when_zoned = [

| `Does_not_exist

| `Invalid_tz_info of string option * Span.t

]

exception Error_when_zoned_exn of error_when_zoned

Constructors

val of_date_and_time : Date.t -> Time.t -> zoneless

Accessors

val date : zoneless -> Date.t

val time : zoneless -> Time.t

Comparison

val equal : zoneless -> zoneless -> bool

Conversion

val to_timestamp_local : zoneless -> timestamp

This yields a "local timestamp" - we pretend we are in the UTC time zone, and calculate seconds since unix epoch

val to_zoned : ?tz:Time_zone.t -> zoneless -> (t, error_when_zoned) Stdlib.result

tz defaults to result of Utils.get_local_tz_for_arg

val to_zoned_exn : ?tz:Time_zone.t -> zoneless -> t

raises Error_when_zoned_exn
if to_zoned fails

val to_zoned_unambiguous : ?tz:Time_zone.t -> offset_from_utc:Span.t -> zoneless -> (t, error_when_zoned) Stdlib.result

val to_zoned_unambiguous_exn : ?tz:Time_zone.t -> offset_from_utc:Span.t -> zoneless -> t

raises Error_when_zoned_exn
if to_zoned_unambiguous fails

val of_zoned : t -> zoneless

Parsing

val of_iso8601 : string -> (zoneless, string) Stdlib.result

Parses a subset of ISO8601, up to 9 fractional digits for second (nanosecond precision).

If more than 9 fractional digits are provided, then only the first 9 digits are used, i.e. no rounding.

val of_iso8601_exn : string -> zoneless

raises ISO8601_parse_exn
if of_iso8601 fails

val maybe_zoneless_of_iso8601 : string -> ([ `Zoned of t | `Zoneless of zoneless ], string) Stdlib.result

val maybe_zoneless_of_iso8601_exn : string -> [ `Zoned of t | `Zoneless of zoneless ]

raises ISO8601_parse_exn
if maybe_zoneless_of_iso8601 fails